ci: green up failing jobs (ws feature, MSRV typo, examples)
ws feature: axum 0.8 changed Message::Text(String) -> Message::Text(Utf8Bytes).
The 7 outgoing send sites in crates/taut-rpc/src/ws.rs now wrap the
serde_json::to_string(...).unwrap() output with .into() so it converts
String -> Utf8Bytes implicitly. Receive side already worked because
&Utf8Bytes coerces to &str via Deref. cargo test --workspace --features ws
now passes 0 failures.

MSRV job: dtolnay/rust-toolchain@1.100 was a typo (Rust 1.100 doesn't
exist). Pinned to 1.75 to match the workspace's rust-version pin.

examples-build:
- counter-sveltekit IncrementInput.by: u64 -> u32 because
  taut_rpc::validate::check::min/max require Into<f64> which u64 doesn't
  satisfy losslessly (the same workaround taut-rpc/src/validate.rs's
  unit tests document). u32 covers the documented 1..=1000 range; the
  saturating_add stays in u64 via u64::from(input.by).
- todo-react CompleteTodoInput now derives Validate (was missing). It
  has no per-field constraints; the empty derive satisfies the macro's
  injected validation hook.

CHANGELOG [Unreleased]:
- Move ws feature from 'broken / v0.2 follow-up' to 'fixed'.

+++ examples/counter-sveltekit/server/src/main.rs
@@ -79,10 +79,11 @@ fn state() -> &'static AppState {
 /// counter to the moon.
 #[derive(Serialize, Deserialize, Type, Validate)]
 pub struct IncrementInput {
-    /// Amount to add to the counter. Must be `1..=1000`. Surfaces as `bigint`
-    /// in TS because `u64` maps to `bigint` per SPEC §3.1.
+    /// Amount to add to the counter. Must be `1..=1000`. Uses `u32` because
+    /// `taut_rpc::validate::check::min`/`max` require `Into<f64>` which `u64`
+    /// can't satisfy losslessly.
     #[taut(min = 1, max = 1000)]
-    pub by: u64,
+    pub by: u32,
 }
 
 // --- procedures --------------------------------------------------------------
@@ -108,7 +109,7 @@ async fn current() -> u64 {
 async fn increment(input: IncrementInput) -> u64 {
     let s = state();
     let mut v = s.value.lock().await;
-    *v = v.saturating_add(input.by);
+    *v = v.saturating_add(u64::from(input.by));
     let new = *v;
     // `send` returns `Err` only when there are zero subscribers; that's the
     // common case before any client connects, so we ignore the result.
